Finally got the interrupts working on A3
This took much longer than I would have liked, but the interrupts are
now finally working on A3, and they appear to be firing everytime now
without fail.

The system is now reading the temperature from the mcp9808 sensor IC,
and sending it to the gateway every 5 seconds.

The signal strength is being updated based on the RSSI return value in
the ACK (from the gateway).